WoT S20 (part 2): rename renderer/ slice folder to render/
Pure folder rename (renderer/* internals untouched, same depth so their imports
are unchanged); only the two external importers updated (table.ts, index.ts).

test:walkontable 745/0/2, test:types + eslint green.
