+++ handsontable/src/3rdparty/walkontable/src/render/_base.ts
