+++ handsontable/src/3rdparty/walkontable/src/render/cells.ts
