+++ handsontable/src/3rdparty/walkontable/src/render/colGroup.ts
