+++ handsontable/src/3rdparty/walkontable/src/render/columnHeaderRows.ts
