+++ handsontable/src/3rdparty/walkontable/src/render/columnHeaders.ts
