+++ handsontable/src/3rdparty/walkontable/src/render/rowHeaders.ts
